# api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/coingecko/tickers": {
            "get": {
                "description": "get Tickers",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "coingecko"
                ],
                "summary": "All Tickers",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/coingecko.TickerRes"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httputil.BadRequestError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httputil.InternalServerError"
                        }
                    }
                }
            }
        },
        "/coingecko/tickers/{id}": {
            "get": {
                "description": "get Ticker by Id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "coingecko"
                ],
                "summary": "Get a ticker",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Ticker ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/coingecko.TickerRes"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httputil.BadRequestError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httputil.InternalServerError"
                        }
                    }
                }
            }
        },
        "/coinmarketcap/tickers": {
            "get": {
                "description": "get Tickers",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "coinmarketcap"
                ],
                "summary": "All Tickers",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/coinmarketcap.TickersRes"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httputil.BadRequestError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httputil.InternalServerError"
                        }
                    }
                }
            }
        },
        "/coinmarketcap/tickers/{id}": {
            "get": {
                "description": "get Ticker by Id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "coinmarketcap"
                ],
                "summary": "Get a ticker",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Ticker ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/coinmarketcap.TickerRes"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httputil.BadRequestError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httputil.InternalServerError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "coingecko.TickerRes": {
            "type": "object",
            "properties": {
                "base_currency": {
                    "type": "string"
                },
                "base_volume": {
                    "type": "number"
                },
                "last_price": {
                    "type": "number"
                },
                "liquidity_in_usd": {
                    "type": "number"
                },
                "pool_id": {
                    "type": "string"
                },
                "target_currency": {
                    "type": "string"
                },
                "target_volume": {
                    "type": "number"
                },
                "ticker_id": {
                    "type": "string"
                }
            }
        },
        "coinmarketcap.TickerRes": {
            "type": "object",
            "properties": {
                "base_id": {
                    "type": "string"
                },
                "base_name": {
                    "type": "string"
                },
                "base_symbol": {
                    "type": "string"
                },
                "base_volume": {
                    "type": "string"
                },
                "last_price": {
                    "type": "string"
                },
                "quote_id": {
                    "type": "string"
                },
                "quote_name": {
                    "type": "string"
                },
                "quote_symbol": {
                    "type": "string"
                },
                "quote_volume": {
                    "type": "string"
                }
            }
        },
        "coinmarketcap.TickersRes": {
            "type": "object",
            "additionalProperties": {
                "$ref": "#/definitions/coinmarketcap.TickerRes"
            }
        },
...
        "httputil.BadRequestError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 400
                },
                "message": {
                    "type": "string",
                    "example": "bad request"
                }
            }
        },
        "httputil.InternalServerError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 500
                },
                "message": {
                    "type": "string",
                    "example": "internal server error"
                }
            }
        },
...
    }
}